feat(client): restructure fallbacks settings into the upstream PluginCard chrome
FallbacksSection -> FallbacksCard (settings.plugin.item): collapsible li
header (name over intro, dirty pill, rotating chevron, aria-expanded) with
the existing form in a divided body, the AC-7 status block folded in, and a
footer with Discard/Reset/Save. Degraded (gateway unreachable) and error
states keep the card chrome with a derived-open notice body (AC-1) plus a
card-local degraded latch so refreshes never collapse the notice; idle loads
config/catalog/switches on mount. Page-only chrome (720px column, title/
intro banners, page-bottom status block) removed; nav locale keys renamed to
title/intro, chrome copy keys added. CSS rewritten on the upstream
PluginCard.module.css vocabulary (--dsw-alias-* only), form fields kept
(h32 r8 / h36 r18). Fidelity checklist filled with verified per-dimension
values. Card spec extended with the chrome contract suite (jsdom +
testing-library); the DOM suite needs a single React instance, so react /
react-dom are linked from the dsh tree by setup-dsh-links.mjs (advisor
pattern) instead of installed as devDependencies.
