+++ package.json
@@ -64,9 +64,12 @@
     "react": "^18.2.0"
   },
   "devDependencies": {
+    "@testing-library/dom": "^10.4.1",
+    "@testing-library/react": "^16.3.2",
     "@types/node": "^22.20.0",
     "@types/react": "~18.3.31",
-    "react": "^18.3.1",
+    "@types/react-dom": "~18.3.7",
+    "jsdom": "29.1.1",
     "schemastery": "^3.18.0",
     "tsdown": "^0.22.2",
     "tsx": "^4.22.4",
